Add base trainer for any accelerate model
